pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "a2wsgi>=1.10.10",
    "email-validator>=2.3.0",
    "fastapi>=0.116.1",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-multipart>=0.0.20",
    "qrcode>=8.2",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
    "websockets>=15.0.1",
]
